+++ src/backends/storage/IStorageBackend.d.ts
@@ -1,6 +1,6 @@
 import * as Koa from 'koa';
 
 export default interface IStorageBackend {
-  middleware(ctx: Koa.Context, next: Function): void;
+  middleware(ctx: Koa.Context, next: () => Promise<any>): void;
   getRelativePath(path: string): string;
 }