+++ src/test/Client.test.ts
@@ -3,18 +3,18 @@ import config from '../config';
 import User from '../models/User';
 import Client from '../models/Client';
 import Project from '../models/Project';
-import {assert} from 'chai';
-import axios, {AxiosInstance} from 'axios';
+import { assert } from 'chai';
+import axios, { AxiosInstance } from 'axios';
 
 const BASE_URL = `http://localhost:${config.PORT}/`;
 
-describe('Client - Integration', function () {
+describe('Client - Integration', () => {
   let client: AxiosInstance;
   let token: string;
   let user1: User;
   let project1: Project;
 
-  before(async function () {
+  before(async () => {
     user1 = await User.create('user1@example.com', 'User1', '123456', true);
     project1 = await Project
       .query()
@@ -41,18 +41,18 @@ describe('Client - Integration', function () {
     });
   });
 
-  after(async function () {
+  after(async () => {
     await Project.deleteAll();
     await User.deleteAll();
   });
 
-  describe('#create', async function () {
+  describe('#create', async () => {
 
-    afterEach(async function () {
+    afterEach(async () => {
       await Client.deleteAll();
     });
 
-    it('creates a client', async function () {
+    it('creates a client', async () => {
       const data = {
         name: 'Test client'
       };
@@ -67,22 +67,22 @@ describe('Client - Integration', function () {
 
   });
 
-  describe('#authenticate', async function () {
+  describe('#authenticate', async () => {
 
-    let client: Client;
+    let authClient: Client;
 
-    before(async function () {
-      client = await Client.create(project1.id, 'Test client');
+    before(async () => {
+      authClient = await Client.create(project1.id, 'Test client');
     });
 
-    afterEach(async function () {
+    afterEach(async () => {
       await Client.deleteAll();
     });
 
-    it('authenticates a client', async function () {
+    it('authenticates a client', async () => {
       const data = {
-        client_id: client.clientId,
-        client_secret: client.clientSecret,
+        client_id: authClient.clientId,
+        client_secret: authClient.clientSecret,
         grant_type: 'client_credentials'
       };
       const response = await axios.post(