+++ src/utils/validate.ts
@@ -1,5 +1,5 @@
-import {get, uniq, union, isPlainObject} from 'lodash';
-import {humanize} from 'underscore.string';
+import { get, uniq, union, isPlainObject } from 'lodash';
+import { humanize } from 'underscore.string';
 import * as validate from 'validate.js';
 import * as moment from 'moment';
 import Media from '../models/Media';
@@ -9,15 +9,17 @@ import Project from '../models/Project';
 
 
 validate.extend(validate.validators.datetime, {
-  parse: function(value: any) {
+  parse(value: any) {
     return +moment.utc(value);
   },
-  format: function(value: any) {
+  format(value: any) {
     return moment.utc(value).format();
   }
 });
 
-validate.validators.lessThanAttribute = function (value: any, options: string, _key: string, attributes: object): string|undefined {
+validate.validators.lessThanAttribute = (
+  value: any, options: string, _key: string, attributes: object
+): string|undefined => {
   if (!validate.isDefined(value)) return;
   const siblingAttributeName = options;
   const siblingAttributeValue = get(attributes, siblingAttributeName);
@@ -28,13 +30,13 @@ validate.validators.lessThanAttribute = function (value: any, options: string, _
   return undefined;
 };
 
-validate.validators.boolean = function (value: any): string|undefined {
+validate.validators.boolean = (value: any): string|undefined => {
   if (!validate.isDefined(value)) return;
   if (!validate.isBoolean(value)) return 'must be true or false';
   return undefined;
 };
 
-validate.validators.arrayLength = function (value: any, options: any): string|undefined {
+validate.validators.arrayLength = (value: any, options: any): string|undefined => {
   if (!validate.isDefined(value)) return;
   if (!validate.isArray(value)) return 'must be an array';
   if ('is' in options && value.length !== options.is) {
@@ -49,42 +51,42 @@ validate.validators.arrayLength = function (value: any, options: any): string|un
   return undefined;
 };
 
-validate.validators.arrayOfIds = function (value: any): string|undefined {
+validate.validators.arrayOfIds = (value: any): string|undefined => {
   if (!validate.isDefined(value)) return;
   if (!validate.isArray(value)) return 'must be an array';
-  for (let id of value) {
+  for (const id of value) {
     if (!validate.isInteger(id) || id < 1) return 'must contain positive integers only';
   }
   return undefined;
 };
 
-validate.validators.arrayOfStrings = function (value: any): string|undefined {
+validate.validators.arrayOfStrings = (value: any): string|undefined => {
   if (!validate.isDefined(value)) return;
   if (!validate.isArray(value)) return 'must be an array';
-  for (let item of value) {
+  for (const item of value) {
     if (!validate.isString(item)) return 'must contain strings only';
   }
   return undefined;
 };
 
-validate.validators.uniqueArray = function (value: any): string|undefined {
+validate.validators.uniqueArray = (value: any): string|undefined => {
   if (!validate.isDefined(value)) return;
   if (!validate.isArray(value)) return 'must be an array';
   if (value.length !== uniq(value).length) return 'must not contain duplicate values';
   return undefined;
 };
 
-validate.validators.choicesUnion = function (value: any, options: {choices: string[]}): string|undefined {
+validate.validators.choicesUnion = (value: any, options: { choices: string[] }): string|undefined => {
   if (!validate.isDefined(value)) return;
   if (!validate.isArray(value)) return 'must be an array';
-  const {choices} = options;
+  const { choices } = options;
   if (choices.length !== union(choices, value).length) return 'contains invalid choice';
   return undefined;
 };
 
-validate.validators.media = async function (value: any, options: {projectId: number}) {
+validate.validators.media = async (value: any, options: { projectId: number }) => {
   if (!validate.isDefined(value)) return;
-  const {projectId} = options;
+  const { projectId } = options;
   if (!validate.isArray(value)) return 'must be an array';
   const ids = [];
   for (const item of value) {
@@ -98,15 +100,15 @@ validate.validators.media = async function (value: any, options: {projectId: num
     .where('projectId', projectId)
     .whereIn('id', ids)
     .first() as any;
-  if (parseInt(result.count) !== ids.length) {
+  if (parseInt(result.count, 10) !== ids.length) {
     return 'contains media not found in project';
   }
   return undefined;
 };
 
-validate.validators.entries = async function (value: any, options: {projectId: number}) {
+validate.validators.entries = async (value: any, options: { projectId: number }) => {
   if (!validate.isDefined(value)) return;
-  const {projectId} = options;
+  const { projectId } = options;
   if (!validate.isArray(value)) return 'must be an array';
   const ids = [];
   for (const item of value) {
@@ -121,16 +123,16 @@ validate.validators.entries = async function (value: any, options: {projectId: n
     .where('entryType.projectId', projectId)
     .whereIn('entry.id', ids)
     .first() as any;
-  if (parseInt(result.count) !== ids.length) {
+  if (parseInt(result.count, 10) !== ids.length) {
     return 'contains entries not found in project';
   }
   return undefined;
 };
 
-validate.validators.projectMember = async function (value: any, options: {project: Project}) {
+validate.validators.projectMember = async (value: any, options: { project: Project }) => {
   if (!validate.isDefined(value)) return;
   if (!validate.isInteger(value) || value < 1) return 'must be a valid id';
-  const {project} = options;
+  const { project } = options;
   const user = await User.getById(value);
   if (!user) return 'invalid user';
   const isMember = await project.isMember(user.id);
@@ -138,7 +140,7 @@ validate.validators.projectMember = async function (value: any, options: {projec
   return undefined;
 };
 
-validate.validators.tags = function (value: any): string|undefined {
+validate.validators.tags = (value: any): string|undefined => {
   if (!validate.isDefined(value)) return;
   if (!validate.isArray(value)) return 'must be an array';
   for (const item of value) {